+minor changes with terminal output +mode fix
